Moved the settings menu from the main activity to the My Store fragment
